requirements.txt
# Core dependencies
psutil>=5.9.0
readline; sys_platform != "win32"

# AI Providers (optional)
ollama>=0.1.0
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
coverage>=7.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Performance monitoring
memory-profiler>=0.60.0